Lab changes 2 may 2013

+++ software/hc.c
@@ -5,7 +5,7 @@ void hc_reset_stream(char rate) {
 	hc_put(0x80000000 | (rate & 0x3));
 }
 
-void hc_preamble(char width, char interval, int offset, int length) {
+void hc_preamble(char width, char interval, short offset, unsigned short length) {
 	unsigned int data;
 	
 	data = width & 0x1;
